fix(postgres): fix bloat size calculation (#1094)

+++ modules/postgres/do_query_bloat.go
@@ -33,11 +33,11 @@ func (p *Postgres) doDBQueryBloat(db *sql.DB) error {
 		case "tablename":
 			table = value
 		case "wastedbytes":
-			tableWasted = parseInt(value)
+			tableWasted = parseFloat(value)
 		case "iname":
 			iname = value
 		case "wastedibytes":
-			idxWasted = parseInt(value)
+			idxWasted = parseFloat(value)
 		}
 		if !rowEnd {
 			return
